fix: fractional crypto qty + real-time refresh + feeds-status panel
User saw three things wrong on the dashboard after the overnight bot's first
crypto fill: position said qty=3 (was actually ~3.99 ETH), cash dropped by
$10K but visible position only accounted for $7K, and the top bar didn't
update. Plus they had Polygon/Finnhub keys in .env but no UI surface
showing what's wired up.

All four problems were real and they share a couple of root causes.

1. Fractional-qty truncation
   ``broker_proxy.get_positions/get_orders`` and ``multi_agent.LivePosition``
   coerced qty to int(). Crypto positions are fractional (3.99 ETH), so the
   missing $2K of position value showed up as "missing cash" — same bug,
   two symptoms. Fixed: keep qty as float through the API; the frontend's
   new ``fmtQty(qty, symbol)`` renders 4 decimals for crypto symbols, 0
   for equities, and "—" for null. Frontend Position type now declares
   numeric fields ``number | null`` to match what the broker proxy
   actually returns for freshly-opened positions.

2. UI never refreshed
   ``TopBar`` queried ``/api/portfolio`` and ``/api/halt`` with no
   ``refetchInterval`` — equity / cash / buying-power numbers stayed
   stale until a hard refresh. Added 5s refetch on portfolio, 10s on
   halt + positions, ``refetchOnWindowFocus: true`` so tab-switching
   forces a fresh pull. Same on ``LivePositionsTable``.

3. .env shadowed by stale shell env
   ``load_dotenv()`` defaults to ``override=False``, so an empty
   ``ANTHROPIC_API_KEY=`` left in the parent shell from a previous session
   blocked the real value in .env from ever being seen. Symptom: bot
   silently fell through to OpenAI / Gemini despite a valid Anthropic
   key. Fixed in ``run_bot.py`` and ``watchdog_agent.py``; also added
   the same load at ``dashboard/api/main.py`` import time so the
   ``/api/feeds/status`` endpoint reflects the user's actual .env.

4. New panel: connected feeds
   ``GET /api/feeds/status`` introspects 13 integration env vars (Alpaca,
   Anthropic, Gemini, OpenAI, Finnhub, Polygon news/stocks/options,
   Quiver, Nansen, Coinbase, Discord) and returns ``configured: bool``
   plus a 4-char tail preview when set. NEVER returns key material —
   verified by ``test_read_status_never_leaks_full_key``. Also rejects
   whitespace-only values so a stale ``export FOO=`` doesn't show as
   green. New ``FeedsStatusBar.tsx`` chip strip on the Portfolio
   page groups by category (broker / llm / news / data / altdata /
   alerts), refetches every 30s, tooltip shows what features depend
   on each key.

Tests: 1229 pass (was 1219; +10 feeds_status). Ruff + tsc clean.

Note for the user: this commit landed while the overnight bot was still
running on stale code. The bot's broker submissions are fine (those were
correct in Round 8); the dashboard reads are what's broken. Restart
uvicorn to pick up the broker_proxy + main.py changes; bot can keep
running in the background uninterrupted.

+++ dashboard/api/multi_agent.py
@@ -73,7 +73,10 @@ class PortfolioEquityResponse(BaseModel):
 
 class LivePosition(BaseModel):
     symbol: str
-    qty: int
+    # qty is float, NOT int — crypto positions hold fractional units (you can
+    # buy 3.99 ETH; truncating to int drops $2K+ of position value and breaks
+    # the cash math reconciliation in the dashboard top-bar).
+    qty: float
     avg_entry_price: float
     current_price: float
     unrealized_pl: float
@@ -467,7 +470,7 @@ async def positions_live() -> list[LivePosition]:
             out.append(
                 LivePosition(
                     symbol=p["symbol"],
-                    qty=int(p.get("qty", 0)),
+                    qty=float(p.get("qty", 0.0)),
                     avg_entry_price=entry,
                     current_price=current,
                     unrealized_pl=float(p.get("unrealized_pl", 0.0)),
